add logger abstraction layer
